requirements.txt
amqp==2.1.4
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
billiard==3.5.0.2
celery==4.1.0
certifi==2017.4.17
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
coverage==4.4.1
coveralls==1.1
daphne==1.3.0
Django==1.11.4
django-bootstrap-breadcrumbs==0.8.2
django-bootstrap-form==3.3
django-bootstrap3==9.0.0
django-bower==5.2.0
django-tables2==1.10.0
djangorestframework==3.6.3
docopt==0.6.2
docxtpl==0.3.9
gitdb2==2.0.2
GitPython==2.1.5
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
kombu==4.0.2
lxml==3.8.0
Markdown==2.6.8
mollie-api-python==1.2.0
msgpack-python==0.4.8
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pycryptodome==3.4.6
PyGithub==1.35
pylatex==1.2.1
PyJWT==1.5.2
python-docx==0.8.6
pytz==2017.2
redis==2.10.5
requests==2.18.1
six==1.10.0
smmap2==2.0.3
Twisted==17.5.0
txaio==2.8.0
urllib3==1.22
vine==1.1.4
zope.interface==4.4.2
